executor: bump per-call timeout
We've started seeing lots of vmalloc failures with the 200ms timeout.
Turns out vmalloc has a check for fatal signals pending,
so we were killing the process which caused vmalloc to fail.
If vmalloc can take 200+ms, then we need to dump the timeout
as it's not blocking that we want to catch.

Bump the timeout to 500ms.