docs: require arm64 build for fuchsia
I am adding a note in the fuchsia docs to specify that the caller should also build fuchsia for arm64. This is because the make extract command will call generate_fidl for TARGETARCH amd64 and arm64. If you happen to have an old tree for arm64, your amd64 changes might be overwritten by the arm64 ones.