tools/syz-testbuild: add utility for kernel build testing
syz-testbuild tests kernel build/boot on releases as it will be done by pkg/bisect.
This allows to ensure that, for example, a change to kernel config won't break
build/boot on older releases and consequently won't break bisection process.
The binary needs to run under root because it creates images.
The kernel checkout given to the tool will be cleaned and used for in-tree builds.
Example invocation:

sudo syz-testbuild -kernel_src $LINUX_CHECKOUT \
	-config dashboard/config/upstream-kasan.config \
	-sysctl dashboard/config/upstream.sysctl \
	-cmdline dashboard/config/upstream-apparmor.cmdline \
	-userspace $WHEEZY_USERSPACE \
	-bisect_bin $BISECT_BIN

A suitable wheezy userspace can be downloaded from:
https://storage.googleapis.com/syzkaller/wheezy.tar.gz
A set of binaries required for bisection (older compilers) can be downloaded from:
https://storage.googleapis.com/syzkaller/bisect_bin.tar.gz