tools/create-image.sh: fix building of i386 images on x86_64 host
When building an i386 image on an x86_64 host, we don't need to use qemu or
check for the presence of an appropriate qemu binfmt_misc configuration.
i386 binaries can run natively, so we also don't need to do debootstrap in
two stages.

Skip qemu checks and run debootstrap in one stage when building i386 on an
x86_64 host.

Signed-off-by: Andrew Donnellan <ajd@linux.ibm.com>