dashboard/config/linux: fix ChromeOS full/baseline configs
ESD_FS/INCREMENTAL_FS were supposed to be enabled in full config only,
but they were enabled only in baseline config.

Split chromeos.yml to chromeos.yml and chromeos-subsystems.yml
to avoid this mistake in future and the need to write [-baseline].

Also enabled UBS_CONFIG_FS as it seems to be enabled on real devices.

Do the same for Android.