vm/qemu: handle QMP events
QEMU occasionally sends events in the same stream used for QMP commands
so from time time the received packet is not a QMP reponse but a QMP event
which breaks the parser. For example, events are send when a machine state
changed.

This adds basic support for event. For now we skip them and wait until
the expected QMP command response arrives.

Signed-off-by: Alexey Kardashevskiy <aik@linux.ibm.com>