syz-manager, syz-fuzzer: filter stale glob values in the corpus
Corpus may accumulate glob values that are already filtered out
by descriptions (e.g. some harmful files), for an example see:
https://groups.google.com/g/syzkaller-bugs/c/W_R0O4XWpfY/m/sdwwg2_hAwAJ

Pass glob files to the manager and filter out values that
are not present in the glob already.

Also use the same caching scheme we use for features and
enabled syscalls so that fuzzers don't need to scan globs every time.