pkg/report: decompile opcodes from reports
Improve Linux reports quality by decompiling "Code: " descriptions.

As that line of opcodes is not guaranteed to begin at the boundary of an
instruction, try to find the right boundary.

Handle the cases of multiple "Code: ..." lines by only decompiling the
first one. In most cases the last such line shows user-space bytes,
which is usually not of great importance.