dashboard/config/linux: disable -Werror on clang instances
Werror is causing headaches for various compiler versions, and Clang
seems to be quite severely hit:
https://lkml.kernel.org/r/YTbOs13waorzamZ6@Ryzen-9-3900X.localdomain

syzbot's resources are better spent on fuzzing, and having the odd
warning in some subsystem penalize fuzzing of the entire kernel is not
appropriate. We assume build-focused CIs will report the warnings (now
errors). If a warning causes a runtime crash, syzbot may even prove that
a warning is a genuine indicator of a runtime problem.

Currently syzbot's KASAN+clang instances are broken due to Werror:

 | arch/x86/kvm/emulate.c:3303:5: error: stack frame size of 2552 bytes in function 'emulator_task_switch'
 | fs/quota/quota.c:766:12: error: stack frame size of 3416 bytes in function 'do_quotactl'
 | sound/core/pcm_native.c:3211:12: error: stack frame size of 2072 bytes in function 'snd_pcm_common_ioctl'
 | net/ethtool/ioctl.c:2698:5: error: stack frame size of 3160 bytes in function 'dev_ethtool'
 | sound/pci/hda/patch_ca0132.c:7026:12: error: stack frame size of 2744 bytes in function 'ca0132_build_controls'
 | drivers/block/loop.c:1818:12: error: stack frame size of 2648 bytes in function 'lo_ioctl'
 | sound/usb/mixer_s1810c.c:543:5: error: stack frame size of 2104 bytes in function 'snd_sc1810_init_mixer'
 | net/ipv6/ipv6_sockglue.c:992:5: error: stack frame size of 2200 bytes in function 'ipv6_setsockopt'
 | drivers/acpi/processor_thermal.c:225:1: error: stack frame size of 2776 bytes in function 'processor_set_cur_state'

Part of the problem is the ongoing issue with CONFIG_KASAN_STACK and
clang: https://bugs.llvm.org/show_bug.cgi?id=38809

There is no easy fix for this in sight.