pkg/csource: fix call list filtering not being consistent
There is a bug in the current implementation that leads to csource using
the original and the new call lists at the same time.

That has led to a bunch of TestGenerate failures.

Enforce the module only to use variables put into the csource context in
order to avoid similar mistakes in the future.