tools/syz-testbed: align checkout stats to the same time
If `max_instances` is smaller than the total number of checkouts in the
testbed configuration file, the averaged per checkout statistics was
inconsistend with respect to the point in time. Checkouts with running
instances were winded back to the current uptime, while others used to
show the complete data.

Fix that inconsistency.