all: add the `DoubleExecCollide` strategy
Add a strategy that resembles the previous collide mode, but detaches not
every other call, rather all calls during the second execution (or at least
as much as possible). Follow the strategy for 33% of all collide
executions.

It was shown during the experiments that this strategy has a positive
effect on the number of discovered crashes and bugs.