tools/syz-check: use $ as struct variant suffix delimiter
We already use this $ convention for syscall variant names.
Use the same convention for struct.
Currently syz-check supports '_' for structs, but it's
inconsistent with syscalls and leads to ambiguity.
If we enable the same matching for all structs (not just netlink),
then '_' creates lots of false matches. E.g. bpf_link_get_next_id_args
is matches with internal bpf_link struct.