dashboard: implement asset storage
Asset is a file/attachment relevant for debugging a kernel crash or the
syzbot itself.

Dashboard keeps track of the uploaded assets and manages their lifetime.
Some of the assets are attached to the bug reports sent by syzkaller,
some (like html coverage report) are displayed on the web portal.

Two new API requests:
* add_build_assets -- let dashboard remember one more build-related
  asset.
* needed_assets -- query the list of assets that are still needed by the
  dashboard.