tools/syz-imagegen: update `erofs` image generation
`./syz-imagegen -fs erofs` fails for some images with the following message:
  <E> erofs: failed to initialize compressor: [Error 22] Invalid argument
    <E> erofs: 	Could not format the device : [Error 22] Invalid argument
This error appears only for the runs with `lz4` compression enabled.
The commit below seems to be responsible for the changed behaviour:
  96a59540090c6376553ff6cd5a042efea53d2d50
  in repo https://git.kernel.org/pub/scm/linux/kernel/git/xiang/erofs-utils.git
Previously, it seems any compression level specified for `lz4` would be
ignored. Now, it is considered incompatible.

This commit removes the compression level. Images will be regenerated in
a future commit.