pkg/symbolizer: match kernel symbol size calculation
Function sizes reported by the Linux kernel do not match symbol tables.
The kernel computes size of a symbol based on the start of the next symbol.
We need to do the same to match kernel sizes to be able to find the right
symbol across multiple symbols with the same name.