dashboard: infer bug info from subject and mailing list
Currently we only determine the bug by looking at the hash encoded after
'+' in syzbot's email address. Users tend to forget to Cc the right
email address while interacting with the bot. Also, it turns out that if
the bug was only sent to Google Groups, any reply to the email thread
does not include the topic starter email address.

Instead of only relying on the right address, try to infer the bug that
was meant by looking more closely at the incoming email.

Parse the mail subject to extract thee title and the sequence number of
the bug.
Extract the mailing list through which the bug was routed to syzbot.
Fetch and filter matching bugs.
If only one matches, handle the command as if it the bug were explicitly
specified.