pkg/image: make Decompress easier to use
Change DecompressWriter to DecompressCheck: checking validity
of the image is the only useful use of DecompressWriter.

Change Decompress to MustDecompress which does not return an error.
We check validity during program deserialization, so all other
uses already panic on errors.
Also add dtor return value in preparation for subsequent changes.