vm/vmm: adapt to recent vmctl changes
The flags now must precede the VM id. The IP address can no longer be
determined by looking at the VM output which was previously possible
due to dhclient printing it. Now that we don't have dhclient we
instead look up the VM id in vmctl status output and the VM IP
uses a fixed pattern.

With the two changes I can run syz-manager and it gets far enough to
report missing coverage.