pkg/asset: support deprecation in the many buckets case
If several syz-cis with different GCS buckets for assets are connected
to a single dashboard, we currently face problems during the asset
deprecation process.

If we receive from the dashboard a valid GCS URL that belong to an
unknown bucket, don't abort the process. Just ignore the URL.

Test this behavior.