pkg/report: gather Go runtime bugs under one bug
When a "fatal error:" bug is reported, this usually means that syzkaller
itself had a memory corruption (except in the gVisor fuzzing case where
this could be an actual bug in gVisor)

Most likely, this is due to a kernel that went wild and corrupted the
syzkaller address space, but in that case the exact details of what part
of the runtime failed are rarely relevant.

This gathers all these go runtime errors under one umbrella so they are
easier to track. Except for gVisor on which the logic is kept the same
as existing.

Add three test cases to the linux reporting:
- 705 (equivalent to the current all/report/7) to make sure Go OOO are
  suppressed (they have a different title now but still get suppressed)
- 706 (equivalent to the current all/report/8) to make sure that ALSA
  "fatal errors" are not handled as Go fatal errors
- 707 (new) to make sure that reports like
  https://syzkaller.appspot.com/bug?extid=3f00d7083c52713ba3b0 are
  re-named to "go runtime error"