pkg/bisect: consider bug's existence
If there's a merge from a branch that was based on a much older
revision, it's likely that it does not YET contain the bug. If we don't
consider this possibility, we're likely to get invalid fix bisection
results for bugs that existed only for a short period of time.

For every fix bisection step, determine first whether the guilty
revision is reachable from it. Cache resuls to speed up processing.

See #4117.