dashboard: generate more cross tree bisections
Currently these jobs are heavily underrepresented -- we rarely get to
generate them and, when we do, we only look up 3 random bugs.

At the same time, for the majority of issues, the check is rather cheap,
we don't even have to query anything from the DB.

Split the job generation results into expensive (= extra DB queries
needed) and non-expensive (= it was enough to examine *Bug) one. Limit
only the amount of the former ones.