sys/syz-sysgen: fail on undefined consts
Ensure that every const mentioned in the description is defined for at
least one architecture.

We cannot do it inside pkg/compiler as it deals with only one arch at a
time. Also, we cannot do it inside syz-check as it's not supposed to be
called every time and may also include non-actionable warnings.

So let's do it inside syz-sysgen, which is invoked all the time and has
all the necessary data.