prog: more complex tests for getInputResources
This commit adds more complex unit tests to cover the bug in
getInputResources fixed by the previous commit.

required_res1 and test_args3 covers the case where a struct is included
both as optional and required. required_res2 and test_args4 cover the
case where a struct is included both as DirOut and DirIn. In both cases
the resource should be recognized as being a required input resource for
the syscall.

Signed-off-by: Paul Chaignon <paul.chaignon@gmail.com>