sys/linux: add null check BPF instructions
This commit adds a new struct with two BPF instructions to perform a
null check on a given pointer. It is then used to update our small ringbuf
program to null check the ringbuf reserved data pointer as follows.

    u64 *e;
    e = bpf_ringbuf_reserve(&rb, sizeof(*e), 0);
    if (!e) return 0;
    [...]
    bpf_ringbuf_submit(e, 0);
    return 0;

With this null check, our test case corresponding to this program now
passes the verifier and is successfully loaded.

Signed-off-by: Paul Chaignon <paul.chaignon@gmail.com>