compiler: support flags as int first argument
This commit adds support for the following syntax:

    int_flags = 1, 5, 8, 9
    int32[int_flags]

which is equivalent to:

    int_flags = 1, 5, 8, 9
    flags[int_flags, int32]

The second int type argument, align, is not allowed if the first
argument is a flag. The compiler will also error if the first argument
appears to be a flag (is ident and has no colon), but can't be found in
the map of flags.

Signed-off-by: Paul Chaignon <paul.chaignon@gmail.com>