all: record diverted bug reproductions
In some cases, we derail during bug reproduction and end up finding and
reporting a reproducer for another issue. It causes no problems since
it's bucketed using the new title, but it's difficult to trace such
situations - on the original bug page, there are no failed reproduction
logs and on the new bug page it looks as if we intended to find a
reproducer for the bug with the new title.

Let's record bug reproduction logs in this case, so that we'd also see
a failed bug reproduction attempt on the original bug page.